requirements/prod.txt
Django==4.2.*
django-betterforms==2.0.0
django-cleanup==8.0.0
django-money==3.4.1
gunicorn
python-dotenv
django-plans==1.0.6
django-recaptcha==4.0.0
psycopg2-binary
email-normalize==2.0.0
Pillow==10.1.0
polib==1.2.0
pyjwt==2.8.0
python-magic-bin; sys_platform == "win32"
python-magic; sys_platform == "linux"
redis==5.0.1
sorl.thumbnail==12.10.0
requests
